Moved swagger to its own router
