+++ src/index.js
@@ -4,9 +4,11 @@ import swaggerJsdoc from 'swagger-jsdoc';
 import morgan from 'morgan';
 import RouteNotFoundError from './error-handler/route-not-found-error.js';
 import {server} from './config/index.js';
+import errorHandler from './error-handler/index.js';
+
 import phoneRouter from './routers/v1/phones/index.js';
 import compareRouter from './routers/v1/phones/compare/index.js';
-import errorHandler from './error-handler/index.js';
+import apiDocsRouter from './routers/api-docs/index.js';
 
 const app = express();
 app.use(express.json());
@@ -15,19 +17,7 @@ app.disable('x-powered-by');
 // APIs
 app.use('/v1/phones', phoneRouter);
 app.use('/v1/phones/compare', compareRouter);
-
-// Standard Stuff
-const openapispec = swaggerJsdoc({
-	swaggerDefinition: {
-		openapi: '3.0.0',
-		info: {
-			title: 'Phone Compare',
-			version: server.version
-		}
-	},
-	apis: ['./src/routers/**/*.js', './src/error-handler/*.js']
-});
-app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(openapispec));
+app.use('/api-docs', apiDocsRouter);
 
 // eslint-disable-next-line no-unused-vars
 app.use((req, res, next) => {
