+++ src/routers/api-docs/index.js
@@ -1,0 +1,20 @@
+import swaggerUi from 'swagger-ui-express';
+import express from 'express';
+import swaggerJsdoc from 'swagger-jsdoc';
+import {server} from '../../config/index.js';
+
+const router = express.Router();
+
+const openapispec = swaggerJsdoc({
+	swaggerDefinition: {
+		openapi: '3.0.0',
+		info: {
+			title: 'Phone Compare',
+			version: server.version
+		}
+	},
+	apis: ['./src/routers/**/*.js', './src/error-handler/*.js']
+});
+router.use('/', swaggerUi.serve, swaggerUi.setup(openapispec));
+
+export default router;
